+++ remote/mdb_client_test.go
@@ -9,7 +9,6 @@ import (
 	"github.com/mongodb/jasper"
 	"github.com/mongodb/jasper/options"
 	"github.com/mongodb/jasper/testutil"
-	"github.com/pkg/errors"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -103,60 +102,6 @@ func TestWireManager(t *testing.T) {
 					tctx, cancel := context.WithTimeout(ctx, testutil.RPCTestTimeout)
 					defer cancel()
 					test.Case(tctx, t, factory(tctx, t))
-				})
-			}
-		})
-	}
-}
-
-func TestWireProcess(t *testing.T) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	makeProc := func(ctx context.Context, opts *options.Create) (jasper.Process, error) {
-		mngr, err := jasper.NewSynchronizedManager(false)
-		if err != nil {
-			return nil, errors.WithStack(err)
-		}
-
-		client, err := makeTestServiceAndClient(ctx, mngr)
-		if err != nil {
-			return nil, errors.WithStack(err)
-		}
-
-		return client.CreateProcess(ctx, opts)
-	}
-
-	for _, modify := range []struct {
-		Name    string
-		Options testutil.OptsModify
-	}{
-		{
-			Name: "Blocking",
-			Options: func(opts *options.Create) {
-				opts.Implementation = options.ProcessImplementationBlocking
-			},
-		},
-		{
-			Name: "Basic",
-			Options: func(opts *options.Create) {
-				opts.Implementation = options.ProcessImplementationBasic
-			},
-		},
-		{
-			Name:    "Default",
-			Options: func(opts *options.Create) {},
-		},
-	} {
-		t.Run(modify.Name, func(t *testing.T) {
-			for _, test := range AddBasicProcessTests() {
-				t.Run(test.Name, func(t *testing.T) {
-					tctx, cancel := context.WithTimeout(ctx, testutil.RPCTestTimeout)
-					defer cancel()
-
-					opts := &options.Create{Args: []string{"ls"}}
-					modify.Options(opts)
-					test.Case(tctx, t, opts, makeProc)
 				})
 			}
 		})
