+++ remote/process_test.go
@@ -1,0 +1,86 @@
+package remote
+
+import (
+	"context"
+	"fmt"
+	"testing"
+
+	"github.com/mongodb/jasper"
+	"github.com/mongodb/jasper/options"
+	"github.com/mongodb/jasper/testutil"
+	"github.com/pkg/errors"
+)
+
+func TestProcessImplementations(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	httpClient := testutil.GetHTTPClient()
+	defer testutil.PutHTTPClient(httpClient)
+
+	for cname, makeProc := range map[string]jasper.ProcessConstructor{
+		"REST": func(ctx context.Context, opts *options.Create) (jasper.Process, error) {
+			_, port, err := startRESTService(ctx, httpClient)
+			if err != nil {
+				return nil, errors.WithStack(err)
+			}
+
+			client := &restClient{
+				prefix: fmt.Sprintf("http://localhost:%d/jasper/v1", port),
+				client: httpClient,
+			}
+
+			return client.CreateProcess(ctx, opts)
+		},
+		"MDB": func(ctx context.Context, opts *options.Create) (jasper.Process, error) {
+			mngr, err := jasper.NewSynchronizedManager(false)
+			if err != nil {
+				return nil, errors.WithStack(err)
+			}
+
+			client, err := makeTestServiceAndClient(ctx, mngr)
+			if err != nil {
+				return nil, errors.WithStack(err)
+			}
+
+			return client.CreateProcess(ctx, opts)
+		},
+	} {
+		t.Run(cname, func(t *testing.T) {
+			for _, modify := range []struct {
+				Name    string
+				Options testutil.OptsModify
+			}{
+				{
+					Name: "Blocking",
+					Options: func(opts *options.Create) {
+						opts.Implementation = options.ProcessImplementationBlocking
+					},
+				},
+				{
+					Name: "Basic",
+					Options: func(opts *options.Create) {
+						opts.Implementation = options.ProcessImplementationBasic
+					},
+				},
+				{
+					Name:    "Default",
+					Options: func(opts *options.Create) {},
+				},
+			} {
+				t.Run(modify.Name, func(t *testing.T) {
+					for _, test := range AddBasicProcessTests() {
+						t.Run(test.Name, func(t *testing.T) {
+							tctx, cancel := context.WithTimeout(ctx, testutil.RPCTestTimeout)
+							defer cancel()
+
+							opts := &options.Create{Args: []string{"ls"}}
+							modify.Options(opts)
+							test.Case(tctx, t, opts, makeProc)
+						})
+					}
+				})
+			}
+		})
+	}
+}
